% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Getters.R
\docType{methods}
\name{classes}
\alias{classes}
\alias{classes,pim.environment-method}
\alias{compare}
\alias{compare,pim.environment-method}
\alias{compare,pim.poset-method}
\alias{names,}
\alias{names,pim.environment-method}
\alias{names,pim.poset-method}
\title{Extract information from pim.environment and pim.poset objects}
\usage{
classes(x)

\S4method{classes}{pim.environment}(x)

\S4method{names}{pim.environment}(x)

\S4method{names}{pim.poset}(x)

compare(x)

\S4method{compare}{pim.environment}(x)

\S4method{compare}{pim.poset}(x)
}
\arguments{
\item{x}{an object of class \code{pim.environment} or \code{pim.poset}}
}
\value{
\code{classes()}: A named vector with the classes of the data
contained in the \code{pim.environment}

\code{names()}: For an object of class \code{pim.environment} the names
of the variables in the object. For an object of class \code{pim.poset},
the name of the poset functions inside the environment

\code{compare()}: A character value indicating how the comparison
is defined in a \code{pim.poset} object, or the poset-slot of
a \code{pim.environment} object respectively.
}
\description{
These functions serve to extract the information contained
in the objects of class \code{\link{pim.environment}} and
\code{\link{pim.poset}}.
}
\examples{
data(DysData)
DysPimEnv <- new.pim.env(DysData,poset=TRUE)
classes(DysPimEnv)
names(DysPimEnv)
compare(DysPimEnv)
}
\seealso{
\code{\link{nobs}}, \code{\link{poset}}, \code{\link{is.complete}}
}

