% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef.R
\docType{methods}
\name{coef}
\alias{coef}
\alias{coef,pim-method}
\alias{coef,pim.summary-method}
\title{Extract the coefficients from a pim or pim.summary object}
\usage{
coef(object, ...)

\S4method{coef}{pim}(object, ...)

\S4method{coef}{pim.summary}(object, ...)
}
\arguments{
\item{object}{a \code{pim} or \code{pim.summary} object}
}
\value{
a named vector with the coefficients.
}
\description{
This function works like \code{\link[stats]{coef}} from the \code{stats}
package. It extracts the coefficients from the objects.
}

