\docType{data}
\name{Engeldata}
\alias{Engeldata}
\title{This is the engel data}
\description{
  This is the engel data
}
\details{
  These are the columns and their meanings \itemize{
  \item{\code{foodexp}} Food expenditure (FE)
  \item{\code{income}} Household income (HI) }
}
\keyword{data}
\keyword{datasets}
\keyword{engel}

