% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Estimators.R
\name{estimators}
\alias{estimator.glm}
\alias{estimator.nleqslv}
\alias{estimators}
\title{Estimator functions for probabilistic index models}
\usage{
estimator.nleqslv(x, y, start = rep(0, ncol(x)), link = "probit", ...)

estimator.glm(x, y, start = rep(0, ncol(x)), link = "probit", ...)
}
\arguments{
\item{x}{a model matrix for the respective pim model. See also
\code{\link{model.matrix}}.}

\item{y}{a vector with the response for the respective pim model.}

\item{start}{a vector as long as there are columns in \code{x}, containing
the starting values for the algorithm}

\item{link}{a character vector describing the link function. ADD MORE INFO}

\item{...}{extra arguments passed down to the actual solver function. See details.}
}
\description{
This page documents different possibilities for solving the score function
of a probabilistic index model or \code{\link{pim}}. All functions mentioned on this page, are essentially wrappers around different solver functions.
}
\details{
All functions share the same three arguments, being the design matrix \code{x},
the response vector \code{y} and the start values for the estimating function.
If you follow the same principles, you can write your own wrapper function
for any solver function of your choice.
}

