% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formula.R
\docType{methods}
\name{formula}
\alias{formula}
\alias{formula,pim-method}
\alias{formula,pim.formula-method}
\title{Getters for a pim object}
\usage{
formula(x, ...)

\S4method{formula}{pim}(x, orig = FALSE, ...)

\S4method{formula}{pim.formula}(x, ...)
}
\arguments{
\item{x}{a \code{pim} or \code{pim.formula} object}

\item{orig}{a locigal value indicating whether the original formula
(\code{TRUE}) or the \code{pim.formula} object (\code{FALSE}) should
be returned. Defaults to \code{FALSE}}
}
\description{
This function allows you to extract a formula from a \code{\link{pim}}
or a \code{\link{pim.formula}} object. In the latter case, you extract
the original formula.
}

