% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Getters_pim.formula.R
\docType{methods}
\name{has.specials}
\alias{has.specials}
\alias{has.specials,pim.formula-method}
\alias{lhs}
\alias{lhs,}
\alias{lhs,pim.formula-method}
\alias{penv}
\alias{penv,pim.formula-method}
\alias{terms}
\alias{terms,}
\title{Extract information from a pim.formula object}
\usage{
has.specials(x)

\S4method{has.specials}{pim.formula}(x)

terms(x, ...)

lhs(x)

\S4method{lhs}{pim.formula}(x)

penv(x)

\S4method{penv}{pim.formula}(x)
}
\value{
\code{has.specials()}: a single \code{TRUE} or \code{FALSE}
value indicating whether the formula right-hand side contains any
special functions.
}
\description{
This group of functions provides an easy way to extract the
extra information saved in a \code{\link{pim.formula}} object.
Take a look at the help page of \code{\link{pim.formula}} for
more information.
}

