% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{new.pim.env}
\alias{new.pim.env}
\alias{new.pim.env,data.frame-method}
\alias{new.pim.env,environment-method}
\alias{new.pim.env,list-method}
\alias{new.pim.env,missing-method}
\title{Constructor for a pim.environment}
\usage{
new.pim.env(data, poset = FALSE, ...)

\S4method{new.pim.env}{missing}(data, poset = FALSE, ...)

\S4method{new.pim.env}{environment}(data, poset = FALSE,
  data.names = ls(data), classes = .get.classes(data), ...)

\S4method{new.pim.env}{list}(data, poset = FALSE, ...)

\S4method{new.pim.env}{data.frame}(data, poset = FALSE, ...)
}
\arguments{
\item{data}{a data frame, a list or an environment containing
the data for a probabilistic index model.}

\item{poset}{either a logical value indicating whether or not a poset
should be added, or list with 2 elements that give the
left- and right hand side of the poset. See also \code{\link{create.poset}}}

\item{...}{extra parameters for construction of the poset and
specification of the parent environment of the object. See also
\code{\link{parent.env}}}

\item{data.names}{An optional character vector with the names of the variables
in the pim environment.}

\item{classes}{An optional character vector with the classes of the
variables in the environment, given in the same order as
the argument \code{data.names}.}

\item{env}{an environment that is the parent environment of the object.}
}
\value{
an object of the class \code{\link{pim.environment}}
}
\description{
This functions serves as a constructor for an object of the class
\code{\link{pim.environment}}. FUNCTION TO BE COMPLETED
}
\section{Methods (by class)}{
\itemize{
\item \code{missing}: 

\item \code{environment}: 

\item \code{list}: 

\item \code{data.frame}: 
}}
\examples{
new.pim.env() # Creates an empty object
}

