% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.pim.formula.R
\docType{methods}
\name{new.pim.formula}
\alias{new.pim.formula}
\alias{new.pim.formula,formula,ANY-method}
\alias{new.pim.formula,formula,pim.environment-method}
\title{Constructor for pim.formula}
\usage{
new.pim.formula(formula, data, ...)

\S4method{new.pim.formula}{formula,pim.environment}(formula, data)

\S4method{new.pim.formula}{formula,ANY}(formula, data, ...)
}
\arguments{
\item{formula}{a formula object}

\item{data}{either a \code{\link{pim.environment}} object containing the
data for the pim, or an object that can be converted to a 
\code{pim.environment} by \code{\link{new.pim.env}}}

\item{...}{extra arguments to \code{\link{new.pim.env}}}
}
\value{
a \code{\link{pim.formula}} object.
}
\description{
This function reworks a formula to a pim.formula for use
in a probabilistic index model. This function is only meant
to be used internally, so it's not exported.
}

