% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{new.pim.poset}
\alias{new.pim.poset}
\alias{new.pim.poset,character,numeric-method}
\alias{new.pim.poset,list,missing-method}
\alias{new.pim.poset,list,numeric-method}
\alias{new.pim.poset,matrix,missing-method}
\alias{new.pim.poset,matrix,numeric-method}
\title{Create a pim.poset environment}
\usage{
new.pim.poset(compare, nobs, parent = parent.frame(), ...)

\S4method{new.pim.poset}{character,numeric}(compare, nobs,
  parent = parent.frame(), ...)

\S4method{new.pim.poset}{matrix,numeric}(compare, nobs,
  parent = parent.frame(), ...)

\S4method{new.pim.poset}{list,numeric}(compare, nobs, parent = parent.frame(),
  ...)

\S4method{new.pim.poset}{matrix,missing}(compare, nobs,
  parent = parent.frame(), ...)

\S4method{new.pim.poset}{list,missing}(compare, nobs, parent = parent.frame(),
  ...)
}
\arguments{
\item{compare}{A character value, matrix or list indicating how the
poset should be constructed. Defaults to the default value of
\code{\link{create.poset}}. See Details section for more information.}

\item{nobs}{An integer value determining the number of observations
this poset is created for. If compare is not a character value,
the number of observations}

\item{parent}{An environment that serves as the parent for the
\code{pim.poset} environment. By default this is the environment
from which the function is called. Note that for a correct functioning,
the parent environment should be set to the \code{\link{pim.environment}}
this object is part of. This is done automatically by the function
\code{\link{add.poset}}.}

\item{...}{arguments passed to other methods.}
}
\description{
This function allows you to create a \code{\link{pim.poset}} environment
that can be added to a \code{\link{pim.environment}} object. You can
use this function to create a custom poset, but in general it's safer
to use the relevant arguments of the \code{\link{pim}} function.
That way more safety checks are carried out.
}
\details{
A poset (or partially ordered set) in the context of
probabilistic index models is a set of indices that determines
which observations are compared with one another. It is used to
construct the pseudo-observations on which the model is fitted.
You can think of a poset as a "pseudo-observation set".

The most convenient way to use this function, is by specifying
a character value for the argument \code{compare}. The value "unique"
creates a poset in such a way that only unique combinations of two
observations are used in the model. The value "all" creates all
possible L-R combinations between the observations.

If you want to define the poset yourself, you can pass either a matrix
or a list with 2 elements as value for the argument \code{compare}.
Columns of the matrix or elements of the list should either be named
"L" and "R", or be unnamed. When unnamed, the function takes the first
column/element as the left poset, and the second column/element as
the right poset.
}
\section{Methods (by class)}{
\itemize{
\item \code{compare = character,nobs = numeric}: 

\item \code{compare = matrix,nobs = numeric}: 

\item \code{compare = list,nobs = numeric}: 

\item \code{compare = matrix,nobs = missing}: 

\item \code{compare = list,nobs = missing}: 
}}
\examples{
mypos <- new.pim.poset('unique',n=10) # creates empty environment
ls(mypos)
# Using the created poset functions L and R
# Note this is purely as illustration, this makes no sense
# in the context of a pim analysis.
mypos$L(1:10)
mypos$R(1:10)
}

