% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nobs}
\alias{nobs}
\alias{nobs,pim.environment-method}
\alias{nobs,pim.poset-method}
\title{Extract the number of observations}
\usage{
nobs(x)

\S4method{nobs}{pim.environment}(x)

\S4method{nobs}{pim.poset}(x)
}
\arguments{
\item{x}{an object of the class \code{\link{pim.environment}} or \code{\link{pim.poset}}}
}
\value{
an integer with the information in the nobs slot. If the
environment is empty, it returns \code{NA}
}
\description{
This function extracts the number of observations in an object
of class \code{\link{pim.environment}}, or the number of observations
for which a \code{\link{pim.poset}} is constructed
}
\section{Methods (by class)}{
\itemize{
\item \code{pim.environment}: 

\item \code{pim.poset}: 
}}

