% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/P.R
\name{P}
\alias{P}
\alias{PO}
\title{Probability function}
\usage{
P(x)

PO(x, y)
}
\arguments{
\item{x}{for \code{P}, a logical value. For \code{PO} a numeric value.}

\item{y}{a numeric value}
}
\value{
either 0 or 1
}
\description{
This functions transform a comparison or otherwise
logical value to a numeric value for use in a pim.
}
\details{
These functions are constructed purely for notation.
\code{P} is completely equivalent
to \code{\link{as.numeric}}, apart from an extra control
to check whether it actually makes sense to do so.
The function \code{PO} is just short for \code{P(x < y) + 0.5*P(x == y)}
}
\section{Functions}{
\itemize{
\item \code{PO}: 
}}
\examples{
# Check in pim
}

