% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/penv.R
\docType{methods}
\name{penv}
\alias{penv}
\alias{penv,pim-method}
\alias{penv,pim.formula-method}
\title{Extract a pim environment from a model or formula}
\usage{
penv(x)

\S4method{penv}{pim.formula}(x)

\S4method{penv}{pim}(x)
}
\arguments{
\item{x}{either a \code{pim} or a \code{pim.formula} object}
}
\value{
In case of a \code{pim} object, the \code{pim.environment}
contained therein. In case of a \code{pim.formula} object,
the environment itself.
See the help page \code{pim.formula-class}.
}
\description{
This function allows you to extract the
\code{\link{pim.environment}} object from either a \code{pim}
object or a \code{pim.formula} object.
}

