% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pim-class.R
\docType{class}
\name{pim-class}
\alias{pim-class}
\title{Class pim}
\description{
This class contains the fitting information resulting from a call to
\code{\link{pim}}.
}
\section{Slots}{

\describe{
\item{\code{formula}}{The \code{\link{pim.formula}} object used in the fit}

\item{\code{coef}}{a numeric vector with the fitted coefficients}

\item{\code{vcov}}{a numeric matrix containing the variance-covariance matrix
of the fitted coefficients}

\item{\code{penv}}{a \code{\link{pim.environment}} object containing the
data used to fit this}

\item{\code{fitted}}{a numeric vector containing the raw fitted}

\item{\code{link}}{a character vector describing the used link function}

\item{\code{estimators}}{a list with the elements \code{coef} and \code{vcov},
containing either a character value with the name of the used estimator,
or the function itself.}

\item{\code{model.matrix}}{If indicated while calling \code{\link{pim}} the original model matrix. Otherwise an empty matrix with 0 rows and columns.}

\item{\code{response}}{}
}}

