% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Getters_pim.R
\docType{methods}
\name{keep.data}
\alias{fitted}
\alias{fitted,pim-method}
\alias{keep.data}
\title{Getters for slots of a pim object}
\usage{
keep.data(x)

fitted(object, ...)

\S4method{fitted}{pim}(object, ...)
}
\arguments{
\item{x}{a pim object}

\item{object}{a pim object}

\item{...}{arguments passed to other methods. Currently ignored.}
}
\value{
\code{keep.data()}: a single logical value indicating whether
the model matrix and pseudo responses were stored in the
\code{\link{pim}} object.
}
\description{
Getters for slots of a pim object
}

