\docType{package}
\name{pimdevel-package}
\alias{pimdevel-package}
\title{Probabilistic Index Models}
\description{
  Fit a probabilistic index model.
}
\details{
  \tabular{ll}{ Package: \tab pimdevel\cr Type: \tab Package\cr
  Version: \tab 1.1.5.4\cr Date: \tab 2013-08-03\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: R (>= 2.12), Matrix, nleqslv, MASS\cr Suggests:
  glmnet (>= 1.9-3), BB, lqa}
}
\author{
  Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve
  \email{JanR.DeNeve@UGent.be}
}
\references{
  [PENDING]
}
\keyword{package}

