% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pim.fit.prep}
\alias{pim.fit.prep}
\title{Prepare the data for a pim model}
\usage{
pim.fit.prep(formula, data, interpretation = c("difference", "marginal",
  "custom"), compare = c("unique", "all", "custom"), ...)
}
\description{
This function is far from finished, but prepares the model matrix
for the \code{\link{pim.fit}} function.
}

