% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pim.fit.R
\name{pim.fit}
\alias{pim.fit}
\title{Fitter function for a probabilistic index model}
\usage{
pim.fit(x, y, link = "logit", estim = "estimator.nleqslv", start = rep(0,
  ncol(x)), ...)
}
\arguments{
\item{x}{a model matrix with as many rows as \code{y}.}

\item{y}{a vector with the pseudo-responses}

\item{link}{a character vector with a link function}

\item{estim}{a character vector or a function indicating the solver
to be used for estimating the coefficients. By default this is
the function \code{\link[nleqslv]{nleqslv}}. Other possibilities are
given in the help page on \code{\link{estimators}}.}

\item{start}{A numeric vector with the starting values for the fitting
algorithm, if required.}
}
\value{
In all cases, a list with the coefficients
}
\description{
This is the basic computing engine called by \code{\link{pim}}
to get the estimates for the coefficients and the variance-
covariance matrices. This function currently only spits out
these components using the sandwich estimators.
}
\details{
NEEDS WORK
}
\seealso{
\code{\link{model.matrix}} for how to construct a valid model matrix
for a pim, \code{\link{pim}} for the general user interface
}

