% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pim.summary-class.R
\docType{class}
\name{pim.summary-class}
\alias{pim.summary-class}
\title{Class pim.summary}
\description{
This class contains the summary information from a probabilistic index
model , and is created by using the function \code{\link{summary}}
on an object of the \code{\link{pim-class}}.
}
\details{
TO BE FINISHED
}
\seealso{
\code{\link{pim}} for more info on how to construct the model
}

