% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{pimdevel-package}
\alias{pimdevel-package}
\title{Probabilistic Index Models}
\description{
Fit a probabilistic index model. This package is the latest version of the pim package, and will end up being pim 2.0. It's a beta, so don't expect back compatibility with previous versions of pim
}
\details{
\tabular{ll}{ Package: \tab pimdevel\cr Type: \tab Package\cr Version: \tab
1.1.6.0\cr Date: \tab 2014-10-16\cr License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr Depends: R (>= 2.12), Matrix, nleqslv, MASS\cr Suggests: glmnet (>= 1.9-3), BB, lqa}
}
\author{
Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve \email{JanR.DeNeve@UGent.be} Joris Meys \email{Joris.Meys@UGent.be}
}
\references{
[PENDING]
}
\keyword{package}

