\name{plot.pim}
\alias{plot.pim}
\title{Plot very basic G.O.F. for a pim}
\usage{
  \method{plot}{pim}(x, y, G = 10,...)
}
\arguments{
  \item{x}{\code{\link{pim}} object.}

  \item{y}{Ignored currently.}

  \item{G}{Number of groups.}

  \item{\dots}{Ignored currently.}
}
\value{
  Nothing.
}
\description{
  Plot very basic Goodness Of Fit for a pim
}

