% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/poset.R
\docType{methods}
\name{poset}
\alias{poset}
\alias{poset,pim.environment-method}
\alias{poset,pim.poset-method}
\title{Extract the poset as a matrix or list}
\usage{
poset(x, ...)

\S4method{poset}{pim.environment}(x, ...)

\S4method{poset}{pim.poset}(x, as.list = FALSE)
}
\arguments{
\item{x}{an object of class \code{\link{pim.environment}}
or \code{\link{pim.poset}}}

\item{...}{arguments passed to other methods. Currently ignored.}

\item{as.list}{a logical value indicating whether the poset should
be returned as list or as a matrix. Defaults to FALSE, which returns
a matrix}
}
\value{
When x contains a poset, either a matrix or a list
(when \code{as.list} is \code{TRUE}) with the indices that
make up the poset. If there's no poset, the function returns
a missing value.

The returned matrix hax 2 columns, each named after the
respective poset function. In case a list is requested, the function
returns a named list with 2 elements, each element containing the
indices related to the poset function of the same name
}
\description{
This function allows you to extract the poset from either a
\code{\link{pim.environment}} or a \code{\link{pim.poset}} object.
The poset can be extracted as a matrix or a list.
}
\section{Methods (by class)}{
\itemize{
\item \code{pim.environment}: 

\item \code{pim.poset}: 
}}
\examples{
data(DysData)
DysPimEnv <- new.pim.env(DysData, poset=TRUE)
poset(DysPimEnv)
}

