% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.pim}
\alias{print.pim}
\title{Display a PIM}
\usage{
\method{print}{pim}(x,...)
}
\arguments{
\item{x}{\code{\link{pim}} object.}

\item{\dots}{Ignored currently.}
}
\value{
Nothing.
}
\description{
Display a PIM
}

