% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.R
\docType{methods}
\name{response}
\alias{response}
\alias{response,pim-method}
\alias{response,pim.formula-method}
\title{Extract response from a pim.formula}
\usage{
response(object)

\S4method{response}{pim.formula}(object)

\S4method{response}{pim}(object)
}
\arguments{
\item{object}{an object of class pim or pim.formula.}
}
\value{
The response variable with pseudo-observations for
a pim.
}
\description{
This function extracts the response from a
\code{\link{pim.formula}} for use in \code{\link{pim.fit}}.
}
\seealso{
\code{\link{pim-class}} and \code{\link{pim.formula-class}}
for more information on the classes, and \code{\link{pim}},
\code{\link{pim.fit}} and \code{\link{pim.formula}} for more
 information on related functions.
}

