% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{The summary function for the pim class}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{an object of the class pim.}
}
\description{
The function \code{\link[base]{summary}} is a generic function. We provide
a method for objects of the \code{\link{pim-class}}.
}

