% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcov.estimators.R
\name{vcov.estimators}
\alias{vcov.estimators}
\alias{vcov.sandwich}
\alias{vcov.score}
\title{vcov estimators for pim}
\usage{
\method{vcov}{sandwich}(Z, beta, Y, W, ...)
}
\arguments{
\item{Z}{the design matrix}

\item{beta}{a numeric vector with the fitted coefficients}

\item{Y}{a numeric vector with pseudoresponses}

\item{W}{a numeric vector with weights. If weights are not applicable,
set to \code{NULL} (the default)}

\item{...}{arguments passed to downstream methods.}
}
\value{
the variance-covariance matrix
}
\description{
\code{vcov.sandwich} and \code{vcov.score} are two similar estimators
for the variance-covariance matrix of a probabilistic index model.
These functions are meant to be used within a call to \code{\link{pim}}
as a value for the argument \code{vcov}.
}
\note{
You should only use \code{vcov.score} in combination with an
identity link
}

