% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.R
\docType{methods}
\name{vcov}
\alias{vcov}
\alias{vcov,list-method}
\alias{vcov,pim-method}
\title{Methods for vcov}
\usage{
vcov(object, ...)

\S4method{vcov}{pim}(object, ...)

\S4method{vcov}{list}(object, ...)
}
\arguments{
\item{object}{any object.}
}
\value{
the variance-covariance matrix
}
\description{
This package defines an S4 generic for \code{\link[stats]{vcov}} and methods for list and pim classes.
}
\seealso{
\code{\link[stats]{vcov}} in the stats package.
}

