% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.poset.R
\docType{methods}
\name{add.poset}
\alias{add.poset}
\alias{add.poset,pim.environment-method}
\title{Add a poset to a pim.environment object}
\usage{
add.poset(x, ...)

\S4method{add.poset}{pim.environment}(x, overwrite = FALSE, ...)
}
\arguments{
\item{x}{a pim.environment object}

\item{...}{further parameters passed to \code{\link{new.pim.poset}}.
Note that you have to provide at least a value for the argument
\code{nobs}.}

\item{overwrite}{a logical value indicating whether the poset
should be overwritten if it's already present. Defaults to
\code{FALSE} to avoid problems.}
}
\value{
The object with a (new) poset attached.
}
\description{
This function adds a poset to a \code{\link{pim.environment}}
object.
}
\section{Methods (by class)}{
\itemize{
\item \code{pim.environment}: 
}}
\seealso{
\code{\link{new.pim.poset}} for the possible values of the
arguments \code{compare} and \code{nobs}.
}

