% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Estimators.R
\name{estimators}
\alias{estimator.BB}
\alias{estimator.glm}
\alias{estimator.nleqslv}
\alias{estimators}
\title{Estimator functions for probabilistic index models}
\usage{
estimator.nleqslv(x, y, start = rep(0, ncol(x)), link = "logit", ...)

estimator.glm(x, y, start = rep(0, ncol(x)), link = "logit", ...)

estimator.BB(x, y, start = rep(0, ncol(x)), link = "logit",
  control = list(NM = c(FALSE, TRUE), method = c(1, 2, 3)), ...)
}
\arguments{
\item{x}{a model matrix for the respective pim model. See also
\code{\link{model.matrix}}.}

\item{y}{a vector with the response for the respective pim model.}

\item{start}{a vector as long as there are columns in \code{x}, containing
the starting values for the algorithm}

\item{link}{a character vector describing the link function. ADD MORE INFO}

\item{...}{extra arguments passed down to the actual solver function. See details.}
}
\value{
a list with following elements:
 \item{coef}{the estimated coefficients}
 \item{conv}{information on the convergence. This isn't
 implemented yet.}
}
\description{
This page documents different possibilities for solving the score function
of a probabilistic index model or \code{\link{pim}}. All functions mentioned on this page, are essentially wrappers around different solver functions.
}
\details{
All functions share the same three arguments, being the design matrix \code{x},
the response vector \code{y} and the start values for the estimating function.
If you follow the same principles, you can write your own wrapper function
for any solver function of your choice.
}
\seealso{
\code{\link[nleqslv]{nleqslv}}, \code{\link[stats]{glm.fit}},
\code{\link[BB]{BBsolve}} for more information on the fitting
algorithms.
}

