% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Getters_pim.formula.R
\name{has.specials}
\alias{has.specials}
\alias{lhs}
\alias{terms}
\title{Extract information from a pim.formula object}
\usage{
has.specials(x)

terms(x, ...)

lhs(x)
}
\value{
\code{has.specials()}: a single \code{TRUE} or \code{FALSE}
value indicating whether the formula right-hand side contains any
special functions.
}
\description{
This group of functions provides an easy way to extract the
extra information saved in a \code{\link{pim.formula}} object.
Take a look at the help page of \code{\link{pim.formula}} for
more information.
}

