% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.matrix.pim.R
\name{model.matrix}
\alias{model.matrix}
\title{Create a model matrix for a probabilistic index model}
\usage{
model.matrix(object, ...)
}
\arguments{
\item{object}{a \code{\link{pim.formula}} object that contains
the formula necessary for constructing the model matrix.}

\item{data}{}
}
\description{
This function creates a model matrix for use in a probabilistic
index model. This model matrix can be passed to \code{\link{pim.fit}}
}

