% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new.pim.env}
\alias{new.pim.env}
\title{Constructor for a pim.environment}
\usage{
new.pim.env(data, poset = FALSE, ...)
}
\arguments{
\item{data}{a data frame, a list or an environment containing
the data for a probabilistic index model.}

\item{poset}{either a logical value indicating whether or not a poset
should be added, or list with 2 elements that give the
left- and right hand side of the poset. See also \code{\link{create.poset}}}

\item{...}{extra parameters for construction of the poset and
specification of the parent environment of the object. See also
\code{\link{parent.env()}}}

\item{env}{an environment that is the parent environment of the object.}
}
\value{
an object of the class
}
\description{
This functions serves as a constructor for an object of the class
\code{\link{pim.environment}}.
}

