% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new.pim.env.R
\docType{methods}
\name{new.pim.env}
\alias{new.pim.env}
\alias{new.pim.env,data.frame,ANY-method}
\alias{new.pim.env,environment,ANY-method}
\alias{new.pim.env,list,ANY-method}
\alias{new.pim.env,missing,ANY-method}
\alias{new.pim.env,missing,missing-method}
\title{Constructor for a pim.environment}
\usage{
new.pim.env(data, poset = FALSE, ...)

\S4method{new.pim.env}{missing,ANY}(data, poset = FALSE, ...)

\S4method{new.pim.env}{missing,missing}(data, poset = FALSE, ...)

\S4method{new.pim.env}{environment,ANY}(data, poset = FALSE,
  env = parent.frame(), data.names = ls(data),
  classes = .get.classes(data), ...)

\S4method{new.pim.env}{list,ANY}(data, poset = FALSE, ...)

\S4method{new.pim.env}{data.frame,ANY}(data, poset = FALSE, ...)
}
\arguments{
\item{data}{a data frame, a list or an environment containing
the data for a probabilistic index model.}

\item{poset}{either a logical value indicating whether or not a poset
should be added, or otherwise a matrix or list with 2 elements that give the
left- and right hand side of the poset. See also \code{\link{new.pim.poset}}
for more information on how to specify a custom poset}

\item{...}{extra parameters for construction of the poset, like
the argument \code{compare} from \code{\link{new.pim.poset}}.}

\item{env}{an environment that is the parent environment of the object.}

\item{data.names}{An optional character vector with the names of the variables
in the pim environment.}

\item{classes}{An optional character vector with the classes of the
variables in the environment, given in the same order as
the argument \code{data.names}.}
}
\value{
an object of the class \code{\link{pim.environment}}
}
\description{
This functions serves as a constructor for an object of the class
\code{\link{pim.environment}}. In most cases, calling this function directly
is not necessary.
}
\details{
This function is called during the preparation of the model
matrix for a pim. The resulting object is used to evaluate the formula
of a pim, and stores information on how this is done.

TODO : Currently there's no automatic assignment of a parent frame, but this
will be added in the near future.
}
\section{Methods (by class)}{
\itemize{
\item \code{data = missing,poset = ANY}: 

\item \code{data = missing,poset = missing}: 

\item \code{data = environment,poset = ANY}: 

\item \code{data = list,poset = ANY}: 

\item \code{data = data.frame,poset = ANY}: 
}}
\examples{
new.pim.env() # Creates an empty object
}

