% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nobs}
\alias{nobs}
\title{Extract the number of observations}
\usage{
nobs(x)
}
\arguments{
\item{x}{an object of the class \code{\link{pim.environment}} or \code{\link{pim.poset}}}
}
\value{
an integer with the information in the nobs slot. If the
environment is empty, it returns \code{NA}
}
\description{
This function extracts the number of observations in an object
of class \code{\link{pim.environment}}, or the number of observations
for which a \code{\link{pim.poset}} is constructed
}

