% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Getters_pim.R
\name{keep.data}
\alias{keep.data}
\title{Getters for slots of a pim object}
\usage{
keep.data(x)
}
\arguments{
\item{x}{a pim object}
}
\value{
\code{keep.data()}: a single logical value indicating whether
the model matrix and pseudo responses were stored in the
\code{\link{pim}} object.
}
\description{
Getters for slots of a pim object
}

