% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pim.fit.prep}
\alias{pim.fit.prep}
\title{Prepare the data for a pim model}
\usage{
pim.fit.prep(formula, data, model = c("difference", "marginal", "custom"),
  compare = c("unique", "all", "custom"), ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that
can be coerced to that class): A symbolic description of the model
to be fitted. The details of model specification are given under 'Details'.}

\item{data}{an optional data frame, list or environment that contains
the variables in the model. Objects that can be coerced by
\code{\link{as.data.frame}} can be used too.}

\item{model}{a single character value with possible values "difference"
(the default), "marginal", "regular" or "customized". If the formula indicates
a customized model (by the use of \code{\link{L}()} or \code{\link{R}()}),
this parameter is set automatically to "customized"}

\item{compare}{a character vector with a single value that describes how the
model compares observations. It can take the values "unique" or "all". Alternatively you can pass a matrix with two columns. Each row represents the rownumbers in the original data frame that should be compared to eachother. See Details.}

\item{...}{arguments passed to other methods.}
}
\value{
An object that can serve as the input for pim.fit. TO BE DONE
}
\description{
This function is far from finished, but prepares the model matrix
for the \code{\link{pim.fit}} function. It has to be redone completely
}
\section{Note}{
 The user has in general no need for calling this
function. It is called by \code{\link{pim}}, which is the preferred
user interface for fitting a probabilistic index model
}

