% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pim.fit.R
\name{pim.fit}
\alias{pim.fit}
\title{Fitter function for a probabilistic index model}
\usage{
pim.fit(x, y, link, ...)
}
\arguments{
\item{x}{a model matrix of dimensions n*p}

\item{y}{a response vector with n values}

\item{link}{a character vector with a link function}
}
\description{
This is the basic computing engine called by \code{\link{pim}}
to get the estimates for the coefficients and the variance-
covariance matrices. This function currently only spits out
these components using the sandwich estimators.
}
\details{
NEEDS WORK
}

