% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.R
\name{pim}
\alias{pim}
\title{Fitting a Probabilistic Index Model}
\usage{
pim(formula, data, link = c("logit", "probit", "identity"),
  compare = c("unique", "all"), model = c("difference", "marginal",
  "regular", "customized"), na.action = getOption("na.action"),
  weights = NULL, keep.data = FALSE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that
can be coerced to that class): A symbolic description of the model
to be fitted. The details of model specification are given under 'Details'.}

\item{data}{an optional data frame, list or environment that contains
the variables in the model. Objects that can be coerced by
\code{\link{as.data.frame}} can be used too.}

\item{link}{a character vector with a single value that determines the
used link function. Possible values are "logit", "probit" and "identity".
The default is "logit".}

\item{compare}{a character vector with a single value that describes how the 
model compares observations. It can take the values "unique" or "all". Alternatively you can pass a matrix with two columns. Each row represents the rownumbers in the original data frame that should be compared to eachother. See Details.}

\item{model}{a single character value with possible values "difference" 
(the default), "marginal", "regular" or "customized". If the formula indicates
a customized model (by the use of \code{\link{L}()} or \code{\link{R}()}),
this parameter is set automatically to "customized"}

\item{na.action}{the name of a function which indicates what should happen when the data
contains NAs. The default is set by the \code{na.action} setting of
\code{\link{options}}, and is \code{\link{na.fail}} when unset.}

\item{weights}{Currently not implemented.}

\item{keep.data}{a logical value indicating whether the model
matrix should be saved in the object. Defaults to \code{FALSE}. See Details.}

\item{...}{extra parameters sent to \code{\link{pim.fit}}}
}
\value{
An object of class \code{pim}. See \code{\link{pim-class}}
for more information.
}
\description{
This function is the main function to fit a probabilistic index model
or PIM.
}
\details{
A BIT MORE INFO ON FITTING MODELS IS NEEDED

It's possible to store the model matrix and psuedo responses in the
resulting object. By default this is not done 
(\code{keep.data = FALSE}) as this is less burden on the memory and
the \code{\link{pim.formula}} object contains all information to
reconstruct both the model matrix and the pseudo responses. 
If either the model matrix or the pseudo responses are needed for
further calculations, setting \code{keep.data} to \code{TRUE} might
reduce calculation time for these further calculations.
}
\seealso{
\code{\link{pim-class}} for more information on the returned
object, \code{\link{pim.fit}} for more information on the fitting
itself, and --- INSERT GETTERS ---
}

