% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimdevel-package.R
\docType{package}
\name{pimdevel-package}
\alias{pimdevel-package}
\title{Probabilistic Index Models}
\description{
Fit a probabilistic index model. The most recent pim version is
the pim package. this pimdevel beta is deprecated and will be removed from
the R-Forge repo after the next major update to the pim package.
}
\author{
Joris Meys \email{Joris.Meys@UGent.be} Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve \email{JanR.DeNeve@UGent.be}
}
\references{
[PENDING]
}
\keyword{package}

