% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pimdevel-package.R
\docType{package}
\name{pimdevel-package}
\alias{pimdevel-package}
\title{Probabilistic Index Models}
\description{
Fit a probabilistic index model. This package is the latest version of the pim package, and will end up being pim 2.0. It's a beta, so don't expect back compatibility with previous versions of pim. Note the fitting engine isn't fully implemented.
}
\details{
\tabular{ll}{ Package: \tab pimdevel\cr Type: \tab Package\cr Version: \tab
1.9.1.8\cr Date: \tab 2015-06-24\cr License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr Depends: R (>= 3.0)}
}
\author{
Joris Meys \email{Joris.Meys@UGent.be} Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve \email{JanR.DeNeve@UGent.be}
}
\references{
[PENDING]
}
\keyword{package}

