% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimdevel-package.R
\docType{package}
\name{pimdevel-package}
\alias{pimdevel-package}
\title{Probabilistic Index Models}
\description{
Fit a probabilistic index model. This package is the latest
version of the pim package, and will end up being pim 2.0. It's a beta
currently under development. There's a pim fitting engine with
rudimentary functionality, but don't expect any back-compatibility
with the previous versions of pim.
}
\author{
Joris Meys \email{Joris.Meys@UGent.be} Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve \email{JanR.DeNeve@UGent.be}
}
\references{
[PENDING]
}
\keyword{package}

