% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Extract response from a pim.formula}
\usage{
response(formula)
}
\arguments{
\item{formula}{a \code{\link{pim.formula}} object.}
}
\value{
The response variable with pseudo-observations for
a pim.
}
\description{
This function extracts the response from a
\code{\link{pim.formula}} for use in \code{\link{pim.fit}}.
#'
}

